Added password req for Updating and Deleteing data

+++ controllers/cocktail-controllers.js
@@ -2,6 +2,7 @@ const asyncHandler = require("express-async-handler");
 const { body, validationResult } = require("express-validator");
 const Liquor = require("../models/liquor");
 const Cocktail = require("../models/cocktail");
+const adminPassword = require("../admin");
 
 exports.cocktail_list = asyncHandler(async (req, res, next) => {
   const cocktails = await Cocktail.find().populate("liquor").exec();
@@ -75,9 +76,16 @@ exports.cocktail_delete_get = asyncHandler(async (req, res, next) => {
 });
 
 exports.cocktail_delete_post = asyncHandler(async (req, res, next) => {
-  await Cocktail.findByIdAndRemove(req.body.id);
-
-  res.redirect("/cocktails");
+  if (req.body.password != adminPassword) {
+    const nameFix = req.params.name.replaceAll("-", " ");
+    const cocktail = await Cocktail.findOne({
+      name: nameFix,
+    }).exec();
+    res.render("cocktail-delete", { cocktail, error: "Incorrect Password" });
+  } else {
+    await Cocktail.findByIdAndRemove(req.body.id);
+    res.redirect("/cocktails");
+  }
 });
 
 exports.cocktail_update_get = asyncHandler(async (req, res, next) => {
@@ -105,10 +113,20 @@ exports.cocktail_update_post = [
     const nameFix = req.params.name.replaceAll("-", " ");
     const cocktail = await Cocktail.findOne({ name: nameFix }).exec();
 
+    const liquors = await Liquor.find().exec();
+
     if (!errors.isEmpty()) {
       res.render("cocktail-update", {
         cocktail,
         errors: errors.array(),
+        liquors
+      });
+      return;
+    } else if (req.body.password != adminPassword) {
+      res.render("cocktail-update", {
+        cocktail,
+        password: "Incorrect Password",
+        liquors
       });
       return;
     } else {
